requirements.txt
arrow==1.2.3
asgiref==3.7.1
attrs==23.1.0
black==23.3.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coverage==7.2.6
Django==4.2.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
Faker==12.0.1
Flask==2.3.2
geos==0.2.3
idna==3.4
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
lxml==4.9.2
MarkupSafe==2.1.2
mixer==7.2.2
mypy-extensions==1.0.0
numpy==1.24.3
packaging==23.1
pathspec==0.11.1
Pillow==9.5.0
platformdirs==3.5.1
pluggy==1.0.0
proj==0.2.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
pyrsistent==0.19.3
pytest==7.3.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
Werkzeug==2.3.4
